requirements.txt
flask
flask-cors
pillow
transformers
torch
google-generativeai
python-dotenv